backend1/emotion_service/requirements.txt
Flask
gunicorn
numpy
scikit-learn
joblib
python-dotenv